WILoops: Remove WILoop peeling
Do not handle kernels with non-uniform predicates with WILoops
(for now). These used to be handled by "peeling" the first iteration of
the work-item loop to check if the barrier is taken (by all WIs) or not,
but it complicates the parallel region formation and control flow construction
quite a bit without known benchmark benefits. The produced peeled loops
are also not easily vectorizable, thus the performance won't be good
anyhow for SIMD.

For VLIW/ILP the peeling might be useful though, but it still doesn't
justify the complexity add.

Handle cond barriers with a fallback handler at the worst case,
which allows removing the peeling logic. Also the explicit unrolling
code was removed to simplify the WILoops pass a lot.