Fix an illegal barrier tail replication case
In case of BBs that are dominated by the same barrier, we should
include them in the same parallel region as the new joined one,
otherwise control flow breakage might occur: We might cause the other
side of a diverging branch to go to a separate parallel region.

See test_id_dependent_computation.cpp which reproduces such a case due
to the diverging branch after the barrier both of which share the same
exit barrier with the early exiting branch due to control flow getting
merged from all function returns to the second last basic block of the
function (all of which leading to the same exit barrier node).