Remove the work-item AA
This is not used anymore. It used to be useful for VLIW
ILP scheduling but we don't use that metadata format
anymore.  I'd guess the disjoint address space is checked
for AA in upstream LLVM, and if it isn't and matters in
some benchmark, we can improve it in LLVM.