Disable the WI-loop beneficiality check
The check is too simplistic as most kernel loops are restricted by loop
carried dependencies, unsafe floating point arithmetics, etc. and most
often than not WI-loop vectorizes better since the OpenCL C is written
with GPUs/SPMD in mind.

We should invent a better check or just improve the LLVM loop
interchange so it can move the WI-loop outside if seen
beneficial. Getting rid of the additional WI context data might
be problematic in that case, but should be possible.