Automatic 'restrict' annotation
POCL_AUTOMATIC_NOALIAS (on by default)

Attempt to infer at runtime if *all* of the kernel command's pointer
arguments could be marked as restrict arguments even if the programmer
didn't annotate them. This is currently set to all or none of the
pointer arguments and only works with the CPU devices. On by default.

Also small cleanups: Two pocl_mem_management function declarations
were in a wrong header.