Remat: Robustness for AllocaInst handling
Sometimes when rematerializing an alloca, the initializer
could fail to remat and we detected that too late. Now
detects remat-ability early to ensure the rematerialization can
be successful.

This doesn't affect upstream PoCL branch so much as
there are much fewer AllocaInsts due to the optimized
input. I have test cases in the upcoming loopvec branch
which failed and validate this fix.

Also removed the Phi branch as we should not encounter
them at this point any longer.