Attempt to remove allocas (specially those that store arguments) using LLVM mem2reg
In #27 there is an example where the user code has a kernel that calls
another function. This indirection is inlined and WorkitemLoops tries
to remove those allocas that are used only for storing function arguments.

However, in this situation there are two allocas, the first stores the argument
value and the second stores the load of the first:

%arg1.addr = alloca ptr, align 8
store ptr %arg1, ptr %arg1.addr, align 8

%arg1.addr.1 = alloca ptr, align 8
%0 = load ptr, ptr %arg1.addr, align 8
store ptr %0, ptr %arg1.addr.1, align 8

WorkitemLoopsImpl::foldTrivialAllocas tries only one time to replace
values for each alloca that uses function arguments. Therefore
allocas that depend on others are not removed like in this example.

This scenario can be triggered by the user, but also by clang from
https://github.com/llvm/llvm-project/commit/642481a4286c9006958274531ee173b347866c50

This commit tries to fix this issue by adding a mem2reg after kernel
function inlining. Since allocas will be removed by this LLVM pass
WorkitemLoopsImpl::foldTrivialAllocas is no longer needed.

The test added verifies that both kernel functions are vectorized.
POCL_VECTORIZER_FORCE_VECTOR_WIDTH=4 is used because of cost model
sees that vectorization is not profitable.